refs CRASM-3382: Move specialized component tests to centralized location (CRASM-3382)
- Move 2 specialized component test files to src/tests/components/
- Rename .spec.tsx files to .test.tsx for consistency
- Update import paths to reflect new file locations
- Remove old snapshot files and empty __tests__ directories
- All moved tests verified working with correct snapshots generated
